requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
configurate==0.4.1
Flask==0.12.2
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
oauthlib==2.0.4
pymongo==3.5.1
PyYAML==3.12
requests==2.18.4
requests-oauthlib==0.8.0
six==1.10.0
tweepy==3.5.0
urllib3==1.22
Werkzeug==0.12.2
